Update port address
From the 'start-server-and-test' docs: The latest versions of Node and
some web servers listen on host 0.0.0.0 which no longer means localhost.
Thus if you specify just the port number, like :3000, this package will
try http://127.0.0.1:3000 to ping the server. A good practice is to
specify the full URL you would like to ping.